backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
docling
chromadb
sentence-transformers
langchain
httpx
python-dotenv